Unify data access type to int across the project
Standardized the type used for acessing data to int throughout the
project. Previously, the project used both int and long for data access,
which pose maintenance challenges and potential risks. By unifying the
data type, we ensure consistency and reduce type-related errors.

+++ quiz1/linux_list.h
@@ -436,9 +436,8 @@ static inline void list_move_tail(struct list_head *node,
 
 typedef struct {
     struct list_head list;
-    long *value;
+    int *value;
 } element_t;
 
 bool linux_list_construct(struct list_head *, int *);
-// void print_linux_list(struct list_head *);
 void linux_list_free(struct list_head *);
